feat: implement saveAuthenticationRequestObj and checkAuthenticationCode in user service
